Adds another option for atropos PE.
I need to be able to dynamically determine if a sample is SE or PE and
then run atropos accordingly.

Problems:
* You cannot use functions as outputs because you need the patterns
* Dynamic outputs gave me problems on large DAGs

I ended up allowing another attributes to be added to `params.R2`,
because params can take functions. Now the wrapper checkts for
`snakemake.params.R2` before deciding that `out_R2` should be none. Then
I needed to add a PHONY rule to point to the atropos rule so that if R2
is needed it can be generated.

This is kind of a specific case, but maybe it will be generally useful.