Update the adapters according to cutadapt docs
In the documentation here: https://cutadapt.readthedocs.io/en/stable/guide.html#basic-usage
it says Illumina now prefers more specific sequences when cutting the adapters as the previous ones were less specific and present elsewhere in the genome. This is apparently true at least for True-Seq. This is a separate pull request since it is in the actual code and not the test.