Update helpers.Rmd to improve bitr (now bitr_override).
This allows bitr to still continue returning IDs when used in neted.lapply even if one of the nested lists contains an ID not in the orgdb database. Usually that is not a problem but if there are merely 1 or 2 IDs and they don't have a corresponding item in the orgdb, bitr normally throws an error. This renames the function to bitr_override for clarity.